added get_random_list, renamed 3.py to copy_random
